+++ src/Sharpen.Engine/Extensions/InvocationExpressionSyntaxExtensions.cs
@@ -1,16 +1,10 @@
-﻿using System;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
+﻿using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace Sharpen.Engine.Extensions
 {
     internal static class InvocationExpressionSyntaxExtensions
     {
-        public static bool IsInvokedWithinLambdaOrAnonymousMethod(this InvocationExpressionSyntax invocation)
-        {
-            return invocation.FirstAncestorOrSelf<AnonymousFunctionExpressionSyntax>() != null;
-        }
-
-        public static string GetInvokedMethodName(this InvocationExpressionSyntax invocation)
+        public static string GetInvokedMemberName(this InvocationExpressionSyntax invocation)
         {
             if (invocation.Expression == null) return string.Empty;