+++ tests/smoke/CSharp50/AsyncAwait/AwaitTaskInsteadOfCallingTaskResult/TaskCannotBeAwaitedInsteadOfCallingTaskResult.cs
@@ -1,0 +1,19 @@
+﻿// ReSharper disable All
+
+using System.Threading.Tasks;
+
+namespace CSharp50.AsyncAwait.AwaitTaskInsteadOfCallingTaskResult
+{
+    class TaskCannotBeAwaitedInsteadOfCallingTaskResult
+    {
+        void CanBeAwaited()
+        {
+            var task = new Task<int>(() => 0);
+            var result = task.Result;
+            result = task.Result + 100;
+            Method(task.Result);
+        }
+
+        void Method(int i) { }
+    }
+}